pyproject.toml
[tool.poetry]
name = "celai"
version = "0.1.0"
description = ""
authors = ["Alejandro Pirola <alejamp@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
#Runtime
python = "^3.11"
# Dependencies for the project:
redis = ">=4.6.0"
numpy = "^1.26.4"
loguru = ">=0.7.2"
chardet = ">=5.2.0"
marko = "^2.1.0"
chromadb = ">=0.5.0"
python-dotenv = "^1.0.1"
openai = ">=1.30.1"
diskcache = ">=5.6.3"
openai-responses = ">=0.3.2"
halo = ">=0.0.31"
aioredis = ">=2.0.1"
fakeredis = ">=2.23.2"
pybars3 = ">=0.9.7"
beautifulsoup4 = ">=4.12.3"
lxml = ">=5.2.2"
shortuuid = ">=1.0.13"
dictdiffer = ">=0.9.0"
pysbd = ">=0.3.4"
deepgram-sdk = ">=3.2.7"
geopy = ">=2.4.1"

[tool.poetry.group.dev.dependencies]
pytest = ">=8.2.0"
pytest-asyncio = ">=0.23.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
